part1-database-etl/requirements.txt
### 📄 part1-database-etl/requirements.txt

```txt
pandas
mysql-connector-python
